requirements.txt
python-dotenv>=0.19.0
pycryptodome==3.19.0
bcrypt==4.0.1
SQLAlchemy==2.0.23
python-jose[cryptography]==3.3.0
cryptography>=3.4.0
